Приступил к обработке формы
