+++ js/main.js
@@ -1,5 +1,5 @@
 import { addToDOM } from './rendering.js';
 import { rawPhotoData } from './data.js';
 import { fullView } from './fullscreen-pic.js';
-import { fileStorage } from './form-processing.js';
-
+import { fileStorage } from './set-new-image.js';
+// form-processing.js
