+++ package.json
@@ -15,7 +15,7 @@
     "browser-sync": "^2.27.10",
     "eslint": "8.7.0",
     "eslint-config-htmlacademy": "5.0.0",
-    "gh-pages": "^4.0.0",
+    "gh-pages": "^4.0.0"
   },
   "license": "MIT",
   "scripts": {
